Reset the clearAllInProgress flag when the shade is collapsed
The NSSL tracks if the clear-all animation is ongoing for things like preventing to launch another dismiss, updating roundness, updating scrolling behaviour and the visibility of the footer.

Currently this flag is only resetted if a StackStateAnimator generated animation has finished, but there are some scenarios when a clear-all is not finalized by these animations, because they didn't run.
This results in hiding the footer, when it needs to be shown and not letting a manual dismiss to finish (leaving a blank view translated out of the visible area).

This CL resets the clear-all animationn flags, when the shade is collapsed. It is addressing the scenario, when the StackSctateAnimator animations were not started, because the shade is closed by the time we get the new list from the pipeline.

Fixes: 273989305
Bug: 283078415
Test: atest NotificationStackScrollLayoutTest
Test: close the shade while the clear-all animation is running, post
some new notifications and see how they behave

Change-Id: Iaff096f037b45a7675db9bf85a39633d9f1837f5

Former-commit-id: 68b9787f2b2e0f370cc91aba12bb0619b8546db9