Provide mandatory gesture insets from the shell caption
This change let the shell caption provide the mandatory gesture insets.
However, note that, due to the lack of relavant insets support for the
shell insets yet, the insets may be unstable during drag-resizing or
drag-moving. Though as the caption insets is there, it should not be a
big issue, and for long term we can get rid of the hack to let the
insets be stable during drag events. This change doesn't contain the
hack for mandatory gesture insets.

Bug: 197277883
Test: WindowInsetsPolicyTest with CAPTION_ON_SHELL enabled
Change-Id: Ifb482311162ef167b4e2d8a375974b7696f9d28c

Former-commit-id: 37bff76c4a1feb54384c4f976b22a7298f941ab7