Avoid invisible PiP handling input events
When the PiP is faded out for an intermediate orientation state
(swiped recents is portrait, display is landscape), its layer
will still intercept input events. Move it outside screen when
its alpha is 0 so it won't affect input.

Also skip mPipAnimationCallback because the direction is not
entering or exiting PiP. That avoids additional position change
when the animation is finished.

Bug: 283971760
Bug: 262835608
Test: Disable launcher rotation. Launch an app which enter PiP.
      Launch another landscape app and swipe up from bottom
      to enter recents. Drag the area where the PiP was, it
      should be able to scroll the recents.
Change-Id: Ibc0ee8e3e4780fc0f6c14c5a54b25c9b30b235c6

Former-commit-id: 7bdba3204187a67fa354bd2bc677b1b52070524d