Animate clock fonts with fewer discrete steps
Currently we're churning a big amount of memory allocating many
variations of font glyphs as we're transitioning between clock states.
That causes a significant amount of allocations, jank possibilities and
memory bloat.

This reduces the amount of taken font variations to, at most, 30 which
significantly increases the amount of font cache hits in
FontInterpolator and reduces the amount of memory we allocate for
glyph paths.

Visually, the difference isn't noticable.

Bug: 275486055
Test: tested visually on foldable and Cheetah on udc-dev builds
Change-Id: I99dd4108eef34156c4f8f79bbb042703325ed82d

Former-commit-id: bcd39307af99f7e8b4a16f29cbff339a8700c4eb