Consolidate adjacent launch while having a background split pair
Launch with launch_adjacent flag will launch the activity to the
adjacent side of the source task and any of the following trampoline
launches should be launched into the same side of its source task. Even
if the trampoline launch reusing one of the splitting tasks in the
background. This makes sure to evicts old child tasks to prevent
interfering with the new split pair.

Bug: 279720005
Test: atest WMShellUnitTests WMShellFlickerTests
Test: having Chrome in a background split pair, click a link on
      Messages, it composes Chrome - Messages pair properly no matter
      what stage Chrome was placed in the background.
Change-Id: I6a8d0cd0000dd13ffdb21a27c927aec03151aabf

Former-commit-id: c67d535eecfc17050e4994b51ae22d46c345b498