Start 2d animation when view laid out.
In Pattern view, we are starting the appear animation before the view is
completely laid out. When the appear animation completes, the dots in
the grid are not in the true position, so when you trigger a layout, the
dots will move to the true position from the position set by the appear
animation. To address this, post the animation in the message queue of
the view. To avoid a bad transition, set alpha to 0 until runnable in
the queue is triggered.

Fixes: 277037626
Test: Pattern View portrait on tablet. Open pattern view and tapping on
top left most dot.
Test: Unlock the device.

Change-Id: I5b0f80fe823615087f5b60646225c6373529fd56

Former-commit-id: 64af0e1b68bf26533df9bbe634ee83380c4dc0f5