Propagate split's task info changes to window decorations
Window decorations were created for split leafs but were not updated
when the task description or it's focused/visible state changed
because the WindowDecorViewModel only received task info changes
from fullscreen and freeform TaskListeners.
This change forwards #onTaskInfoChanged calls from the stage
TaskListeners to the WindowDecorViewModel so that its internal state
is always up to date and to allow decorations from being correctly
drawn or hidden.

Bug: 280831370
Test: atest WMShellUnitTests
Test: manual - put Gmail and Chrome into split screen, then open a
fullscreen app from the taskbar. Verify that the fullscreen task's
caption responds to drags to enter desktop.

Change-Id: I5ce5df3abf540f7f09e42a70f48cb8b07d64fc99

Former-commit-id: d647128979e5570e59afa52dd6400098144ca3a6