Fix the processgroup creation failure due to race conditions
...between creating the process and attaching the process.

The webview/app zygote process creation consists two steps:
a) Fork the process; b) Create the processgroup.
Betwen a) and b), if the app process has started and attaches
itself to the system_server, and also dies before b), we'd
end up killing its processgroup before creating it. This CL
fixed this issue.

Bug: 283391058
Test: Manual
Change-Id: I2c68f43ac921621bf0a3b0b2b399118518c60ca1

Former-commit-id: 00bf6da8468fdc2b97f626dbad024f78795c5d4b