Update decor insets when the amount of insets source is changed
For example, a display cutout has a top insets with 50px. Then when
adding status bar with 20px, the size of insets is still 50px so no
need to trigger display size change. But for a different rotation,
the display cutout will be at another side, then the top insets will
become 20px from status bar.

The problem will be an additional display config change when the
first time orientation change happens. The case will only happen
once per boot.

This change ensures that the insets in non-current rotation are
updated when the insets source is added or removed. Though the count
is not 100% accurate for any layout param changes, it is the most
efficient way to correct the state only once when initializing
(usually only when booting) the insets windows instead of re-compute
for every layout change.

Fix: 280930112
Test: DisplayPolicyTests#testUpdateDisplayConfigurationByDecor

Change-Id: Ieedcc6388237b7feccd8a869105a5f8e70ff4afd

Former-commit-id: 2f95cf02db412c355d779eb9d7088912c635cb2b