Remove ability for user to open wallet directly within shortcuts disabled dialog
Currently the picker only allows launching an activity upon a disabled shortcut scenario.

In the case that this activity isn't resolvable, this leads to a bad user experience, even if we prevent the crash.

For now, remove the encouragement to open the Wallet directly, and replace it with the default "Done" button. This is how Device Controls handles it as well.

Bug: b/280554185
Test: atest QuickAccessWalletKeyguardQuickAffordanceConfigTest.kt
Change-Id: I7adb8783f31fc3c9563fa84d13cfd9cec42d7f5e

Former-commit-id: cec4e04001279bfea29a47cc7af7bbfb6b4be017