Fix TextAppearanceInfo#getTextColor is wrong when there is a hint text
When there is a hint text, the text color is set to hint color in TextView. While what the input method really want is the normal text color rather than hint color.

We can't replace "textPaint.getColor" with "textView.getCurrentTextColor()" directly, because we need to retrive color from the textPaint we specially constructed above when there is a spannable text that overrides the text color by spans.

Bug: 284088242
Test: atest android.view.inputmethod.TextAppearanceInfoTest#testCreateFromTextView_withHintText
Change-Id: Ib11173a18ddd8c1cac7be1340c65daf6f2ff804a

Former-commit-id: c52cdb5cd6d16f7f52100151ba996042c4a6112e