Set fit insets type for decor insets test
Otherwise because since the key of InsetsState#mSources is based
on hashcode (InsetsSource.createId), the order to iterate mSources
may be indeterministic for the same insets type.
Then WindowLayout#computeFrames may compute different results
according to the hashcode, that causes the test to be flaky.

By setting fit insets type to 0, the insets won't affect each other.

Fix: 284435531
Test: DisplayPolicyTests#testUpdateDisplayConfigurationByDecor

Change-Id: I30ea22c051f84264a03d0b698d50d83c62059869

Former-commit-id: 614ed9e0e99fa7a18d8785e6f6087b667e4e7af4