Attempts to fix InputChannel not initialized crash.
This is a blind fix (as I couldn't reproduce the original crash)
targeting a race condition where setView might be called after the host
is already destroyed.

Fix: 283001352
Test: Manually verified no harm done.
Test: Navigating into and out of WPP and
between the Shortcuts and main screen still seems to work. Using the
quick switch UIs for wallpaper and color on the main screen seems to
still be working as well.

Change-Id: I82864d91d7a35845446806eb705c64ba7a1b949f

Former-commit-id: f5811c4c530c3024186e242485434a9945f073eb