Fixes i18n issue with shortcut names.
Some lock screen shortcut names were not updating to their translated
names when the user changed system language, remaining with the old
language in Settings and in Wallpaper Picker.

This CL fixes that issue.

The cause for the issue was how some pickerName properties were
implemented using a getter and some were constants, without a getter.
The constant ones were not being translated.

The approach in the CL is to actually move from a property to a method
call. This way, there is no choice but to recalculate the name of the
affordance every time; closing the door on the reintroduction of this
issue when a new config implementation is added with the developer
skipping the getter.

Fix: 280077350
Fix: 280829365
Fix: 278858630
Test: changed languages and verified that the shortcut names are updated
to the new language in both "Wallpaper & style" and in Display > Lock
screen

Change-Id: I2d3e3a621b6f24c42c1600ce4af17617dd52560a

Former-commit-id: 18bd504bce6304584e35ddba3a0a5a937eae4c8d