[Unfold animation] Merge change transitions into unfold transition
Currently if we unfold a foldable device, first the display switch
happens and then after some time Launcher adds a taskbar, which leads
to an extra configuration change (and a CHANGE transition).

This causes a problem that unfold animation gets stuck for some time
if this second CHANGE transition starts during unfold animation.
This CL fixes it by merging this extra transition into the unfold
transition.

Bug: 278064943
Test: unfold several times on launcher, app => check that it looks
 correct
Change-Id: Id0fa208210e275042a84a40f6e823652314a306c

Former-commit-id: 46b22adb71cbb0af876ad9de552a69eadfb9a39a