Partial screen sharing: show error notif when screen recording fails
When starting a screen recording of a specific app, and the app is then
removed from recents, there would be an error toast shown, but the
"recording" notification was still showing, and interacting with it
would do nothing.

With this fix, we show an error notification for that case, which
replaces the "recording" notification.

Fixes: 250033241
Test: RecordingServiceTest.java
Test: Manually - 1. Enable partial screen sharing
                 2. Start recording a specific app
                 3. Remove the app from recents before the recording
                    starts
Change-Id: Ic2dc8e6c40ba13d5feccc3deface0f62edabd6c8

Former-commit-id: 588bfe051b45743701304a94d63b1eb43fdf8fb1