Don't cancel notifications when pausing work apps
Previously when the profile was turned off, all apps were killed, so it
made sense to cancel all the notifications after that. If the profile is
kept running, and suspension is used instead, there is no need to cancel
the notifications, because they'll be hidden for suspended apps anyway.
Also, if the notification gets cancelled, it is possible that the user
will never see it, especially when the profile is paused by schedule.
This CL makes the behavior consistent with focus mode - when the profile
gets unpaused, all work notifications reappear, those that existed
before the profile was paused and those that were posted while it was
paused.

Bug: 285129642
Test: manual with Telegram messenger
Change-Id: I45bfa87c36478237d04375a51473b35ec7704e0b

Former-commit-id: 1075d5196b1ac719501f35c1adcf63772759cdb5