Fix lockscreen anim not running on unlock with FPS
For side FPS devices, it is possible for the device to be waking up
and also dreaming. If this happens a
BiometricUnlockController.MODE_WAKE_AND_UNLOCK_FROM_DREAM is
sent. This will then take path that assumes the lockscreen content
will be hidden immediately, as if from AOD. This isn't the case for
side FPS devices, where the animation still needs to run.

Test: atest KeyguardUnlockAnimationControllerTest
Test: manual - This is a timing issue. Unlock with side FPS over and
over again to ensure the animation runs
Fixes: 283672908

Change-Id: Ia5c426596977a4e98ad4abdb65ce9b72d8154045

Former-commit-id: 51341e5c27152ffbab647a7a79ec8f0037e7340e