Ensure embedded window's channel is not disposed before being duped
If the InputChannel is created inside a critical section, but then
duped to another channel after the lock is released, it's possible
the channel could be disposed (e.g. because window is removed) before
the duplication happens. To prevent this, dup the InputChannel in the
same critical section as where it was created.

Bug: 283314325
Test: Presubmit
Change-Id: I00dd3e54240512a6d2ddfe7ec40a6c4f14d837fc

Former-commit-id: 2eb99915cf14409c1b4dd149f7d270ae49d34272