Fix LS Smartspace Text Color
The LS Smartspace was applying the wrong text color because the views
were out of bounds due to incorrect calculations with the actual on-screen
coordinates. This fix now restricts the view to be within the
device width and height.

This fix also addresses an issue where the region samplers would not
persist on the lockscreen smartspace because they were only being
created once and immediately removed when the view was detached from
window but never recreated. Now, the region samplers are recreated when
the view is re-attached so that text protection on the LS smartspace
will persist.

Flag: region_sampling
Fixes: 283620770
Test: manual, atest RegionSamplerTest
Change-Id: Id6bb46c35dcb9e37c9825ba9836a7d5152384844

Former-commit-id: 076f2c6d695e3d657f587a46b6e5347137f7321f