Avoid applying weird wallpaper offset
The case may almost only happen with image wallpaper that will
set requested size larger than screen to support scroll. Usually
it is hard to observe because it may be updated to correct state
in a short time.

Sometimes the window hierarchy may have updated to new orientation,
then the parent frame will be assigned based on configuration bounds.
But the specified requested size has higher priority to affect
window frame. So if the wallpaper client has not updated the new size
according to the new orientation, the offset may put wallpaper at a
weird position.

Bug: 281973565
Test: Rapidly switch between apps and home with different orientation.

Change-Id: I53a44f389e7e8a9ae193e1e4dfac6e0f344e5200

Former-commit-id: 296f8a1e0971bbd56fcfa56a6e1a95cfc56a2c2b