Fix wrong gesture insets size during rotation
We use a pre-provided layout params to deal with rotation. However, the
cutout in the pre-provided layout params is not calculated based on
rotation correctly. Use the minimal insets size in display cutout safe
field instead to let the gesture insets to be with the same size in any
given location to fix the issue.

Besides, the mMinimalInsetsSizeInDisplayCutoutSafe is missing in IPC,
fix it by adding correct statements in the parcel related methods.

Bug: 285090475
Test: Rotate the device with cutout and there's no flakiness.
Change-Id: Iaddf3acf61a80cce356d009d5a3ca7539f1a33a0

Former-commit-id: 0cb9c102249bd118cabe9f2e058f479588e9180b