Don't bring child to top if fullscreen task open
We will try to dismiss split screen if we got not children when
startAnimation, but we always try to bring one side tasks to top
and it might behavior not correct when user try to open another
app.

If transition contain a fullscreen opening change, we should not
bring any side to top to make sure the opening task will show on
top.

Fix: 282894249
Test: Open map to navi and any other app as split, then
      open Settings from notification shade, Setting should on top
Test: pass existing tests
Change-Id: Ia36a3ae8def02471d21efa9194bf67ab63e644a2

Former-commit-id: b0876d564132a1280928c310bdaf7bf819d86769