Hide keyguard earlier when exiting lockscreen
Call hideKeyguardViewAfterRemoteAnimation earlier, when starting to fade
in the home wallpaper. This call will hide the lock wallpaper.

With this, then the lockscreen live wallpaper flag is enabled,
hideKeyguardViewAfterRemoteAnimation will be called another
time from exitKeyguardAndFinishSurfaceBehindRemoteAnimation in
onAnimationEnd. But the second call will always be ignored. Replace
the error log with an info log when the second call is ignored, since
this is expected.

Bug: 283390961
Bug: 284081102
Test: manual, unlock with/without bouncer, with fast/short swipe, with/without AOD, ...
Change-Id: Icfd993cc222ee8fd59d25beecb982cc545104d82

Former-commit-id: 9df43f863e07fd7a9f904b0793d994b108eec087