Fix inline reply defocus animation issues
Instead of adding the RemoteInputView to an Overlay, it is kept normally in the view tree and is thus still receiving windowInsetsAnimation callbacks during the animation.

In order to free up the space occupied by the RemoteInputView during the animation, a negative topMargin is set during the animation.

Bug: 281847513
Flag: NOTIFICATION_INLINE_REPLY_ANIMATION
Test: atest RemoteInputViewTest, Manual (i.e. analysing screenrecordings of inline reply defocus animation and verifying that "clear all" button is functioning normally afterwards)
Change-Id: I8a9e1ab227d204ca18554a447f9f2ccc404be73e

Former-commit-id: 202096373b1083163b9551a8da6a489f47e86d20