Remove CentralSurfaces.isPanelExpanded
CSI.mPanelExpanded was redundant, because the only code path that
updated its value also updated NPVC.mPanelExpanded to that same value.
Additionally, whenever CS updated its value, it would then tell NPVC
to update the sysui state flags using its copy of the data. This change
eliminates CSI's copy and getter, and it moves the flag update request
into NPVC to eliminate some spaghetti code. There was a tiny chance
that there was another listener for NPVC.mPanelExpanded changes that
received the event before CSI and needed the sysui state flags to not
have been updated yet, so I verified that that was not the case.

Bug: 249277686
Test: manual
Change-Id: I05a038d2be66cfb4d84662d0f18465e161564653

Former-commit-id: 35fd336aa4c1a8fb751967a764b973676cff4bc2