Don't let expanding notification shade cancel unlock
At the end of unlock, views start resetting themselves. This
includes the notification shade which goes from 0f height
to 1f height (covering the keyguard again).

Since notification shade height is 1 minus keyguard height,
and the full unlock finalisation happens in a handler to avoid
jank, this can cause a second hit to dismissAmountThresholdsReached
which cancels the unlock.

Test: android.platform.test.scenario.sysui.foldable
Bug: 284096414
Bug: 282672298
Change-Id: Ic9469c015c47aade323710a877923e8b0c8960b9

Former-commit-id: 209d5a532831bb71e1c259b2b3a02bab315e5aa6