[Dumpsys CLI] Remove Exception bounds on LogBufferEulogizer
LogBufferEulogizer previously used `<T : Exception>` as a generic bound
for its `record` method, but that doesn't seem necessary, since any
`Throwable` would work. The method is used in situ when an exception
is being handled, so that the handler can record the exception and
still throw it.

This change relaxes the bounds to `Throwable` and removes the one place
that was filterig out non-exception throwables in SystemUIService.

Test: all sysui tests
Bug: 278094048
Change-Id: I6f7d1309890724c1e2d335602dd9706bac4f5aa4

Former-commit-id: ae48859e88ea4570a1a8fcb0d0fcae8f1a66a172