[pm] disable deprecated ABI warning dialog for tests with abiOverride
The deprecated ABI warning dialog is breaking quite a few UI tests that
are launched in the 32 bit mode. Since Tradefed uses --abi flag to
differentiate the 32-bit and 64-bit test cases, we pass the flag through
ApplicationInfo so that the warning dialog can be skipped if the --abi
override flag has been used. This should fix all the UI tests that are
launched in 32-bit.

Test: atest with broken tests
Test: abtd
BUG: 284553354
BUG: 283266877
BUG: 285587518
Change-Id: I161142dde7c6f6e33650e3d31ee56c01c7de7d6b

Former-commit-id: 155275fcfc812e8969fb67c57c5334f22bdb4687