Only apply transient visibility for root task
For example, assume all task are opaque (from top to bottom):
TaskR(root)
 -TaskA(top split)
   -TaskX
   -TaskY
 -TaskB(bottom split)
   -TaskZ
Only TaskX and TaskZ should be visible. TaskY is invisible
because it is occluded by TaskX.
(The problem was that TaskY becomes visible)

In the other words, when entering recents, TaskR is occluded
by recents so only need to force TaskR to be visible. While
the descendant tasks of TaskR should still follow the original
visibility calculation that computes whether they are occluded
inside the parent task.

Fix: 279128366
Test: atest RootTaskTests#testGetVisibility_MultiLevel

Change-Id: Ied02650f08918471257d1cefda9fae365654fd4c

Former-commit-id: 6008eca456b1a92f3621ff4d8a36c6ec1ca4142c