Avoid unexpected turning screen on when top activity finishes
The ActivityRecord#mCurrentLaunchCanTurnScreenOn is true while
the Activity was started and created. However, another activity
may immediately started afterward. Once the top activity finishes
while the display is off, the display was unexpected turn on
since the next top activity is able to turn screen on.

Reset the can-turn-screen-on state of the last started activity
when starting a new activity.

Bug: 270827073
Bug: 241961202
Test: verified on sample app
Change-Id: I2c2d15e1a1a68d583d42dde69aef2f596cd064b5

Former-commit-id: 2af9efb27589bb1a233a5db1d4da303fd09c32e8