Handle UNKNOWN sim state in KGVM.
In keyguard view mediator, we do not handle the UNKNOWN sim state
correctly. In CentralSurfaces, we are preventing dismiss amounts to be
considered because the sim state is considered secure.  When the esim is disabled, it is confirmed that we expect an
UNKNOWN sim state. This state does not propagate to the view mediator
from the update monitor. When the sim is disabled, the subscription id
is -1 and state is UNKNOWN(0). We want to ensure that this state change
is propgated to view mediator to ensure that the sim state is no longer
considered secure. Additionally, a SIM_PIN_REQUIRED signal is sent as it
transitions from SIM_PIN_REQUIRED to UNKNOWN. This is setting
mPendingPinLock to true. This makes keyguard show after it is dismissed
the first time.

Fixes: 280010618
Test: disable sim and swipe to unlock LS with none security method
Test: unlock sim and swipe to unlock LS with none security method
Test: disable sim and unlock with pin
Test: unlock sim and unlock with pin
Test: disable sim puk and swipe to unlock LS with none security method
Test: unlock sim puk and swipe to unlock LS with none security method
Change-Id: I5bf6ce457a2d87d8f27ca907684a3e1571f9252b

Former-commit-id: fb1db79cf48be8f8e43d32201a0cd2a0ea68f2d7