Fix screen not turning off on power button press.
If Bedtime is active and the "Keep the screen dark" option is enabled,
then the device shouldn't dream and tapping power button should turn the
screen off but didn't. This change fixes that issue.

Bug: 284857728
Test: Manually by enabling bedtime and the "keep screen dark" option and
then tapping power button to verify the screen turns off. Also, if
bedtime is disabled (or "keep screen dark" is disabled), then power
button should continue to start dreaming.

Change-Id: I5db541c5c2e52c58d773822bb7404704516cbb28

Former-commit-id: 9c486ed4fcd2bc28bb009be8b4d80fed6c6532cd