[Decor] Remove unused properties and Tunable
Before adding support for runtime debugging, we can remove the `Tunable`
implementation from ScreenDecorations, as well as the unused fields
declared.

Tunable was added to the screen decor when the assumption was that a
single path was suitable for all rounded corners. Therefore, the only
supported tuning was `SIZE`, which simply changes the size of the
rounded corners. With the proposed changes, this tuning size is no
longer relevant. The CLI that will be implemented can handle arbitrary
paths, viewport sizes, and colors, and does not require registering
entries with the TunerService.

Test: ScreenDecorationsTest
Test: RoundedCornerResDelegateTest
Bug: 285941724
Change-Id: If9f3e28c641590c04d19315970359bed04521bc6

Former-commit-id: 68270b2198c6613bc0df15c6dea0a8ea073efc84