Requesting a new transition when user switches
Home task was moved to top while switching to a new user and
the home activity was started afterward. A new transition was
started while starting the home activity and the previous top
activity visibility changes were also collected. The transition
was somehow aborted since the launch result was DELIVERED_TO_TOP.
Therefore, the previous top activity remained visible in the
background.

Requesting a new transition when user switches, so that all the
hierarchy changes would be collected and the visibility would
be updated after transition completed.

Bug: 284458878
Test: verified by the repro steps on the bug
Change-Id: I27177ba4b0017f89701d82cb4ffbb3db106725ea

Former-commit-id: ddda9bd87aa5bfdbd171a45d07874bc2c2c9f890