Don't dismiss split during enter if already visible
Open to side enter transition is only used if split is active and
foreground. We will check exception case if changes on both stage
didn't include in this transition. But this exception case should
happen usually while user trying to launch same app which already
on front. So we should not dismiss split itself and just log it.

Fix: 285795299
Test: manual
Test: pass existing tests
Change-Id: Idf6c38ade49ce25f1301e0ca4bb17210dcb906f9

Former-commit-id: 0a4215b3c1a0a01c9f66cd0b19649d8717cc7516