[Decor] Make ScreenDecorations.mDebug non-final
Most of the code behind DEBUG was logging-related. This CL converts most
logs to use the LogBuffer implementation, but keeps a few `onPreDraw`
logs behind a new `DEBUG_LOGGING` flag, which is still statically
defined.

This CL also moves `DEBUG` to a new `mDebug` field, which is private and
modifiable at runtime. The `setDebug` method is unused in this CL, but
will be supported in follow-up. Setting debug is still supported through
the old sysprop `debug.screenshot_rounded_corners`. In the future,
setting debug will be supported over the command line, and will
essentially re-setup the screen decor wih the flag enabled.

Test: adb shell setprop debug.screenshot_rounded_corners 1 && restart-sysui
Test: ScreenDecorationsTest
Bug: 285941724
Change-Id: Ibc19ae82257d8e5fa4c45582d7d83067bbf67548

Former-commit-id: 76a732f8e527554cdf7f36bb6eec71c5c6213556