Pass LayoutParams from SCVH to Parent window
There are some params that a view or window can set that needs to be
handled by WMS. However, SCVH have no interface into WMS since they are
handled in the client process. Instead, pass the params to their host
parent and let the host ensure the correct params are handled in WMS.

Test: SurfaceControlViewHostTests
Bug: 281687307

This reverts commit 7b5b92f136fe970e837730118b64c1535d12a6ff [formerly 41dd323ac08077f22b9e305d9be9848a19c564d9].

Reason for revert: Disable failing test. b/286245842

Change-Id: If0e61bfc4d1e32513316d4a9a22faf2ec77e690c

Former-commit-id: 0d73e1abeb557deafa9cc26b3a41673287ee55fa