Camera: release extension when extension app crash
When extension use app crashed, extension proxy service does not clean
up the SessionProcessorImpl or ExtenderImpl. This could cause the
resource leak as all resources allocated is expected to be released from
the onDeInit or deInitSession.

This patch will attached DeathRecipient to the SessionProcessorImpl and
ExtenderImpl. And, it will handle relevant release sequence.

Bug: 286198063
Test: Manual test with the extension test app that intentionally crashes
Change-Id: I9053bc9a51e65c33309eea78db538dfc9962affc

Former-commit-id: b2456b4058803abe865f1109666bae824268f79e