Repackage APK to resolve testCheckSignaturesRotatedAgainstOriginal
All of the APKs that are used by the PackageManagerTests do not have
a valid uses-sdk element, so they are treated as having a
targetSdkVersion of 0. While most APKs were previously on the GPP
allow-list and would not result in a failure during install, the
APK for testCheckSignaturesRotatedAgainstOriginal was not scanned
until recently and was blocked by GPP during install. This commit
repackages this APK to resolve failures from this test.

Fixes: 283797480
Fixes: 286712534
Test: PackageManagerTests
Change-Id: Iaf0e974190273b13468da5e61eb8c8df73607ffc
Merged-In: Iaf0e974190273b13468da5e61eb8c8df73607ffc

Former-commit-id: f4cbf1dd3b5e3a31c2b392a075e2e19229ae57bb