Fix surface didn't fade out when dismiss
We did fade-out on snapshot if exist rather than surface itself
before due to the surface didn't have content. But this issue
seems had been fixed so it cause we cannot see the animation due
to the surface layer is higher than snapshot.

Now we can fade out the surface directly so just remove using
snapshot in this case.

Fix: 286353033
Test: manual
Test: pass existing tests
Change-Id: Icd04940f69b4a4102068be14785718ea8ae77a3a

Former-commit-id: 7a9b53cd31356fc0b3d1d65ad136c891d11e7be3