Set FontInterpolator cache sizes to fit expected animation steps
This right sizes the FontInterpolator caches so that they are never
expected to evict. This prevents us from every creating new fonts
at interpolated positions after the first animation, which hides a
lower level leak in the font rendering framework related to SystemUI
never being put into the background.

I expect this won't prevent leaking entirely as clock changes will
delete this cache and recreate it, thus reinitializing another set
of interpolated font objects. However this is not expected to be a
terribly common operation (unlike animating between the lockscreen
and AOD)

Bug: 275486055
Test: Checked RssAnon manually on local device under stress test
Change-Id: I3f342fd7d427b8b6cfcd2e83480fbbb36d6782ea

Former-commit-id: 02804291ce68767526c75f6b30e853cb15cc8879