Fix screen getting relocked after unlocking once and wallpaper dream has stopped
For windowless dream, onAppTransitionFinishedLocked is not
called when dreaming starts like other dreams, so
mLockAfterAppTransitionFinished flag remains true.
When user unlocked via bouncer and dreaming stopped,
onAppTransitionFinishedLocked got called and lockNow would
then be called to relock the screen.

Fix is to add isDreaming check so when dreaming has stopped,
we won't call lockNow.
Renamed the variable to mLockAfterDreamingTransitionFinished
as it is only used by power button press to start dreaming.

Bug: b/285059622
Test: atest PowerKeyGestureTests; manual
Change-Id: Iceeb2ac97514600bd9a52439282496a38c387b98

Former-commit-id: 6102ee30e65e6e6d308d59f81bed19935a7bc1c7