Demote tryRemoveNotification crash to Log.wtf
We're ending up removing a notification twice -- probably a race
between clear all and dismissing or cancelling one of the notifications
hit by the clear all.

Short term, we don't want to bother our users with a crash, and we think
the end state is okay -- a notification removed twice will be removed the
first time, so the UI state should be okay -- so this changes the "throw
IllegalStateException" in tryRemoveNotification to a Log.wtf, which will
only crash *droidfood* builds.

Bug: 282945192
Test: presubmit
Change-Id: I65d7a9587946c51258263e5ea3069b4b5cdbc82e

Former-commit-id: 89c150dac053efbd332a1321e4c3199a1c5845e9