Remove windowing mode override from task when it matches display
When a task is moving from desktop to fullscreen or vice versa, check if
the desired windowing mode matches the display windowing mode. If it
does, remove the windowing mode override and set it to undefined.

Fixes an issue with split screen where a task that has been on desktop,
is not eligible to be in split anymore. As the task had windowing mode
overridden to fullscreen.

Bug: 285358244
Test: atest DesktopTasksControllerTest
Test: move Chrome to freeform and back to fullscreen, use "new window"
  from the menu to create a new chrome window in split screen
Change-Id: I7b683e1cf2353c5cfa58bd1d7d3ab1b5319c9be7

Former-commit-id: 55997d1ae8c0d770d55f0fc4cd0e38a3dc5af076