Put DWBC calls on handler
Without DWBC calls on a handler, it was possible for DPC.stop() and
therefore dwbc.disable() to be called whilst querying other dwbc data.

This change puts the critical methods on the dpc/dpc2 handler, so that
this issue does not occur.

Bug: 283886342
Test: atest com.android.server.display
Change-Id: Ia0688642189b050353b45dd31431d5f157b428b8
Merged-In: Ia0688642189b050353b45dd31431d5f157b428b8

Former-commit-id: 846833639b7f1c922a63bde69dd7f35fdb26f34b