chore(#MagSettingsPanel): make seekbar index reflect the controlling magnifier scale
Originally the panel has content observer to let seekbar progress sync with Settings persisted scale. With previous cl that we notify WindowMagnification the controlling magnifier scale updated, we can make the seekbar index update with notified new scale. Therefore, the seekbar index could reflect the actual controlling magnifier scale. Also, if the seekbar triggers onProgressChanged with fromUser false, which means the progress changed is not by the user interaction but by the notifying of the magnifier scale changed, we don't need to trigger the callback to change the magnifier scale.

Besides, we remove the persisting scale logic in WindowMagnificationSettings. Since the persisted scale would be updated when the callback onPerformScaleAction is triggered in framework service side, we don't need the duplicated update in sysui side.

Bug: 286340381
Bug: 286342154
Test: manually
      atest IWindowMagnificationConnectionTest
      atest MagnificationSettingsControllerTest
      atest WindowMagnificationControllerTest
      atest WindowMagnificationSettingsTest
      atest MagnificationControllerTest
Change-Id: Ibf78c90bc282225c2cd7670a52cb3a5e686d7d3f
Merged-In: Ibf78c90bc282225c2cd7670a52cb3a5e686d7d3f

Former-commit-id: 96261a5373386831df7f8286096f0d920c29d221