Fix 3-finger swipe unlock not working
FalsingManager is currently rejecting trackpad 3-finger events. The x and y movements in 3-finger motion events are generated from the gesture library according to the movement on the trackpad. The movement scale is bigger than events on the screen, hence considered "fake" by the falsing manager. We should skip all trackpad events in gesture falsing logic (since falsing is an on screen thing to begin with).

Fixes: 277273475
Test: 3-finger swipe to bring up bouncer
Change-Id: I1223129ec40c0ed0e81e5f740564172665284585

Former-commit-id: abdc3384bc81c66c3a4c22881092093738e765e9