Fix IME insets not apply hidden when switching focus in diffrent users
Normally, the focus window will apply the IME visibility state to
WindowManager when the IME has applied it. But it would be too late
when switching IMEs in between different users. (Since the focused IME
will first unbind the service to switch to bind the next user of the
IME service, that wouldn't make the attached IME token validity check
in time)

As a result, we have to notify WM to apply IME visibility before
clearing the binding states in the first place.

Fix: 285821212
Test: manual as issue steps
  1. Launch youtube with personal profile
  2. Launch Gmail with work profile
  3. Split youtube with personal profile and gmail with work profile
   (The Work Profile app is split and shown upper)
  4. Typing something with personal profile app
  5. Click work profile app at blank space. (Non input box)
  6. Expects IME insets should be hidden without seeing the black area
Test: atest DefaultImeVisibilityApplierTest#\
              testApplyImeVisibility_hideImeWhenUnbinding
Change-Id: I0873ec065722e9e0b22dd5010d3415d3704cfd2d

Former-commit-id: 56fa64663f909bf79367c3c2122737a6c7a8c0d1