Fix ScreenOffBrightnessControllerTest
This test was failing on devices where the configuration for
config_allowAutoBrightnessWhileDozing defaulted or was set to true.

The config should be true, when allowing full automatic screen
brightness when in doze, and therefore, screenoffbrightnesssensor should
be enabled, only when this value is false, since it is used as a
fallback sensor.

The timings of the setting the return value of this testable resource
needs to come before the variable is actually set when creating the DPC,
which is why the DPC needed to be re-initialised in this test.

Bug: 286860986
Test: atest DisplayPowerController2Test
Change-Id: Ieea7b056f6858e606b5039735e3f0ba729c94646

Former-commit-id: b31a5b62343d240a7161fbb89f2b1cbd1d4484d7