chore(non linear font scaling): remove mAlwaysScaleIcon in StatusBarIconView
The constructor StatusBarIconView(Context context, AttributeSet attrs) to set mAlwaysScaleIcon be true is never used, so mAlwaysScaleIcon is always false in StatusBarIconView lifetime. Therefore, we remove mAlwaysScaleIcon and the constructor so that we don't need to consider mAlwaysScaleIcon value in future cl.

Besides, we add private method isNotification() to wrap the check whether mNotification exists.

Bug: 282111042
Test: pass the build
Change-Id: I5eaea511718bd70681840948b54b2f178cf0e85e

Former-commit-id: c7f002fd681cc62d9a415e17b0eeb83679c69e1d