Add TrustedPresentationCallback API for AttachedSurfaceControls
The TrustedPresentationCallback is a public API but need to be
registered on the SurfaceControl that is rendering content. This means
that windows can't register a TPC because their rendering SC is not
exposed. Add an additional API that allows the TPC to be called on
AttachedSurfaceControl so the listener can be registered on the
rendering SC.

Test: TrustedPresentationCallbackTest#testSetTrustedPresentationListenerOnWindow
Bug: 278027319
Bug: 287076178
Bug: 279957280
Change-Id: I9c6cc4745c0dcbbabc0477ada2faba9ca597a9d7

Former-commit-id: 52976ae302d004cecf665d6819b072ba615406e1