Don't transfer snapshot starting window if bounds has changed.
The snapshot starting window could just remove itself when receive
resized request, so transfer it to a different size activity is
meanless but only cause flicker. So when the trampoline activity
finish and attemp to transfer the snapshot starting window to the
real activity, skip transfer starting window at that time, the
transaction will ready when app window drawn.
And in order to prevent the trampoline activity trigger an activity
transition animation, reuse the no_animation flag, so that transiton
won't merged into playing transition.

Bug: 278931413
Test: hot launch test app on large screen device, verify no flicker
from transfer snapshot starting window.

Change-Id: I285eccac8cea15846ca6c36d987d077bd0edde8d

Former-commit-id: bbc669b2476e9cd40c854d5633faf43ae871987e