Fix crash opening avatar picker in non-en-US locale.
This fixes a crash seen when opening the avatar picker if the current
language is not en-US.

The crash occurs because content descriptions are provided in a sring
array in an overlay, but the base empty array has been translated and
is replacing the actual descriptions in non-default locales.

This CL adds translatable=false to the base array, and adds a check
to not crash if the content descriptions array is smaller than the
images array.

Bug: 284422901
Test: Manual, try opening avatar picker in various locales.
Change-Id: I4a518cd37c4ca7139d182612cd02d79489948474

Former-commit-id: f756b2623fa76b9d0c3ca2ee9f2894856bb1e448