Force rebind live wallpapers when destination change
When a live wallpaper is on system-only, and the same wallpaper
component is reapplied to system+lock, the binding would be skipped due
to the "changingToSame(...)" check. Given that the picker will call
clear(FLAG_LOCK) after applying the wallpaper to system+lock, and given
the lockscreen wallpaper current implementation, this would lead to a
black screen. Forcing the rebind prevents the black screen issue.

Also, log the "which" flag in setWallpaperComponent since it's often
useful.

Test: atest WallpaperManagerTest
Test: manual: follow the repro steps of b/286428230 with/without this CL
Bug: 286428230
Change-Id: I7f70f676ae0c4eb480e3049af49d2b5232d9360b

Former-commit-id: 7b1c22cde1cf54d8f4124a9d6aac5a2c528a2330