Move ImmersiveModeConfirmation related resources to sysUi (5/n)
This CL copy the resources related to immersive mode confirmation window
to the system UI side to make sure changes won't affect existing server
side window.

Compared to the original resources, the system UI side resource removes
all AccentColor and will set it in the code side. To achieve that,
additional id is also given. Added neccesary `android:` desciption to
some values.

This should be submitted together with, or before the window code.

Test: build
Bug: 277290737
Change-Id: Id463ae98c495c09897826068bdc27dad30c33447

Former-commit-id: 819df2ef684010436400f774ccf8b380c965f14c