Fix status bar is barely visible when the dim window is above it
Previously, the system will force the status bar content to white when
the dim window is shown. It's ok for the case when the dim window is
below status bar window but it will make the content barely visible if
the status bar background of an app is also white and the dim window is
above the status bar window.

We should seperate the two cases and only update the color when the dim
window is below status bar window.

Bug: 241274551
Test: 1. Add an account and setup fingerprint unlock
      2. Go to "Settings>Passwords & accounts>add account" to add
	 another account
      3. Wait for BiometricPrompt to pop up and check the status bar.
Change-Id: Ib4e2cbb34b6fecd50da58dc018bd39e309dea47b

Former-commit-id: 8f5412527d76b5daddc4b86802c080b633b372ab