Avoid refreshing the media output dialog UI while dismissing.
refresh() currently performs several binder calls on the UI thread.
During launch of the dialog this happens before the animation, so while
there is a slight delay there are no frame drops. However when closing
the dialog the call happens after the animation has already started,
causing it to stutter significantly.

The proper fix would be to rearchitect this dialog so that state can be
loaded ahead of time and in the background, so the UI can start off
pre-populated and updates can come in whenever available. This would be
a risky change at this point in the U release cycle, and I'm not
familiar with the code enough to write it confidently, so for now simply
avoiding the updates if the dialog is dismissing removes the jank,
though it is only a band-aid.

Bug: 287191450
Flag: N/A
Test: see videos in the bug
Change-Id: If25937c86499b947a2f0d9971783f153f8e1a743

Former-commit-id: 103188a89c2e143d93a22a8e55f5ead46ee4b7ad