Update EmptyShadeView during GONE->AOD transition
Internal changes to keyguard have caused our call to
NSSLC#updateShowEmptyShadeView to happen *before* keyguard states have
been updated for the GONE->AOD transition. This is due to an implicit
assumption that by the time our unrelated call to
updateShowEmptyShadeView occurs, the transition state is already set.

This is fixed by explcitly listening for the GONE->AOD transition and
calling updateShowEmptyShadeView when it starts and ends.

Fixes: 286209517
Test: atest NotificationStackScrollLayoutControllerTest

Change-Id: Ie81ea7f95fa42294fb32cffd305f840ca9c1c9e3

Former-commit-id: cc70765f6af2a154565d49353642e6ed0f8d0a7c