Reparent PiP overlay under task as soon as possible
For enter PiP animations started from Launcher, we send the signal to
WMShell to reparent the overlay under task as soon as possible.

This avoids flicker issue caused by race condition of cleaning up the
transition leash before the reparenting was done, which made the issue
of overlay being not attached to anything for few frames.

Bug: 281711939
Test: winscope trace from comment#28, leash follows task
Change-Id: I46764c925185c316b3aa309e3cb9f1a0a89948d4

Former-commit-id: 8aa7dcd12db070abea7912d5f8a0c31b471378ba