Hide animated ghost when underlying view is detached from window.
This mirrors the behavior that we have when the view is made
INVISIBLE/GONE. The need for this change is that a view can be detached
with its visibility remaining VISIBLE, in which case the updated
location of the animated ghost become (0,0), which makes the transition
look jumpy and broken.

Note that this also disables dismiss animations for the
MediaOutputDialog when the "Stop casting" button is tapped, since we
know that the notification will go away but can't rely on the timing of
it. Disabling the animation makes the dialog fade away in place instead,
so it has a reliable transition.

Fix: 232542592
Flag: N/A
Test: see before/after videos in the bug
Change-Id: Ieba6a2f5681ad7143c2f14b59f0309f658d66ce9

Former-commit-id: afa8d3986898f20c2f459668b3082a68f137fc8b