Do not add divider to dismiss transition
If there is also a display change in same transition, we need to
release old divider surface and it lead to crash due to it running
animation before release.

Actually, we cannot see divider animation in dismiss transition
due to it didn't be added in transition root so its animation
is invisible. So remove this didn't affect current dismiss animation.

Fix: 287207156
Test: manual
Test: pass existing tests
Change-Id: I9c0cacfbd8df7deed5c2b028c1e2af9a0cac540a

Former-commit-id: c6feebfd9598e106f0cb1394271da14b5b9cb983