Fix privacy chip overlapping with statusbar icons
According to my investigation there was a race condition on setting the alpha value of the end-side statusbar content. It could happen that the "show animation" for the end-side statusbar content was triggered. If at roughly the same time a statusbar chip animation is started, it could happen, that the chip animation is finished before the "show animation", which causes the "show animation" to set the final alpha value, even though the chip animation should take precedence (since the statusbar icons should always be hidden if there is a chip currently being animated).

Bug: 287923736
Test: atest MultiSourceMinAlphaControllerTest, atest CollapsedStatusBarFragmentTest
Change-Id: Ifa16bf3aff784c98063b09064571a30a51030bb0

Former-commit-id: 01572146e07d88cc8df280acc2840d8840f7db3d