[CS] Define an #alwaysCollectFlow method that can be used from Java.
As we move more CentralSurfacesImpl methods to flows, Java classes will
need to listen to them. Some of these Java classes need to be *always*
listening to flows and cannot tie the flow collection to a specific
view. This CL adds a helper method for Java classes to do that.
ag/23710783 will use this helper method.

As we move more of our codebase to recommended architecture, I hope this
method will be used less frequently and can eventually be deleted.

Bug: 277764509
Test: compiles
Change-Id: Iff8eba81597626941b34b2fdc1b96d0acd9ca5e5

Former-commit-id: 954c40f0f84028fda7289eab96a243cee08d2a02