Fix how taskview is used by controlsUi
When controlsUi is dismissed, it calls taskview#release, which
should only be done when the task & taskview are done being used
which isn't really the case until the task is explicitly removed.

Since the controlsUi is going away, the taskview surface gets
destroyed which triggers a transition to move its task TO_BACK,
however, because release has been called, we have no reference to
the taskview in TaskViewTransitions, so the TO_BACK transition
falls to the default transition handler which results in the
incorrect animation.

To fix this, instead of calling taskview#release when the controls
ui is being dismissed, call taskview#removeTask to actually remove
the task. In the onTaskRemovalStarted callback, we can then clean
up the taskview.

Test: manual - open controls panel and swipe up to home
             => observe that there isn't a second animation of the
                task animating away
Bug: 282102652
Change-Id: Ie40d078e4a05d3323a3364b82f6af171ae97181c

Former-commit-id: 96c2b131230db84221c23c07ac4853ed02fc6461