Fixes bad PIP state when opening Maps from the lock screen
When opening an app that's in PIP from a notification on the lock screen,
there's a race condition where the EXIT_PIP transition occurs before the
keyguard is removed, causing it to get aborted. This change resets the
PIP state and windowing mode manually when we receive the aborted transition.

Also fixes an issue with a snapshot being present when entering PIP
from this state.

Test: Manual check + presubmit
Bug: 283783868
Change-Id: I026dd35498210d2b39ccc0235ec225a9c7341dfe

Former-commit-id: 2cca90ecf47eac6d465ee6d105154286cce7dfdb