chore(#MagSettingsPanel): change persisting scale timing
Originally the Settings presisted scale is updated when the panel seekbar progress is changed or the window magnifier scale is changed by accessibility actions. In WindowMagnificationController we keep the same behavior. In WindowMagnificationSettings, per b/286176069 we change to only update the persisted scale with the seekbar final progress when the user interaction on seekbar is ended. We change the onPerformScaleAction parameters so when sysui call the service to change the magnifier scale, we can decide if the persisted scale should be updated.

Bug: 286176069
Test: manually
      atest WindowMagnificationTest
      atest MagnificationSettingsControllerTest
      atest WindowMagnificationSettingsTest
      atest WindowMagnificationControllerTest
      atest WindowMagnificationManagerTest
      atest MagnificationControllerTest
Change-Id: I995a059c6575298eeae095fb3daa9f037916014a
Merged-In: I995a059c6575298eeae095fb3daa9f037916014a

Former-commit-id: 58b977cfd1dcd2b9d631cb1e12212cfd1f8e7bfe