Do not make device is asleep a gating condition for face auth
This prevents face auth from running when the user touches the UDFPS sensor on AOD
The device would still be asleep but we still want to run face auth.

This also fixes race conditions between device asleep still being true when we trigger face auth by lifting the device or tapping the device to wake up.

Fixes: 285499709
Fixes: 285501230

Test: atest DeviceEntryFaceAuthRepositoryTest
Change-Id: Ieaef83a2695de1c2d369068b4db08df5acb2357d

Former-commit-id: 37b214b8c61342f99c342bc3e68b953ddf452f74