Docs: Grammar fix (it's -> its)
The docs linter throws a warning for this -- several uses of the phrase
"it's contents are backed" and it should be "its contents are backed"
("it's" means "it is" or "it has"). Checked the other uses of "it's" in
this doc and they're all fine.

NO_TEST , fixes a doc bug (no code edits).

Change-Id: Id29a4ddb25c1f3cfc4001aab12785d44cd409c1a
TEST: [go/abtd docs build]
FIXES: 288932046

Former-commit-id: 444b464ca61984b598123ae910abc7936d443d61