Don't set ImageWallpaper to home only the first time
CL only effective if lockscreen live wallpaper are not enabled.

The first time, setting a wallpaper on home only will
trigger a migration of the default wallpaper, which causes some
problems when the lockscreen live wallpaper flag is not enabled.

In that case, force the wallpaper to be set on home + lock, to make sure
the lock screen doesn't end up without any wallpaper.

Bug: b/279129369
Test: manually check that the bug is fixed
Test: atest WallpaperManagerTest
Change-Id: Id955a8336455b47a8ba30958f25f2fd276fee013
Merged-In: Id955a8336455b47a8ba30958f25f2fd276fee013

Former-commit-id: 7a5705b3c2106dd1fa2c0484c8a9ff6a92509f1e