Use passed in Transaction when creating bounds layer in VRI
When the bounds layer is created, the visibility and initial bounds are set
using a new transaction that is applied immediately. This can cause a
race condition because the initial bounds are sent using a different
apply token than the buffer updates. When the crop for the bounds layer is
updated, that is sent in sync with the BBQ frame which has a separate
apply token. This means the updated bounds can be applied before the
intial bounds, causing unexpected behavior.

Test: testCropWithChildBoundingInsets 100 times
Bug: 288339794
Change-Id: I61776050e99d8fa9524a7b6e4fe6034c171d49a7

Former-commit-id: 215c18cfc83c2a07e27e9f2e7d1556b272f9e88d