Play insets animation while forcibly showing system bars
Before this CL, when a system bar is being forcibly shown, InsetsPolicy
would set the control target to null, which just resets the leash
without any animation. This CL sets the proper control target and uses
the InsetsController to play insets animations.

This CL also applies the same logic to the transient bars of immersive
mode.

Fix: 283738201
Test: InsetsPolicyTest
Test: Open an app from kid space, and see if navigation bar is shown
      with animation.
Test: Send a Chat message to a device which is in immersive mode, and
      see if status bar is shown with animation.
Change-Id: I8178a9cabfa62a3942ca6ff51b31df22c0d570d3

Former-commit-id: 4203281253f89b91aab954a70a5b9bad8e6592ce