[CS] 3/N: Move KeyguardViewMediator to WallpaperRepository.
Note: KeyguardViewMediator gets recreated on user switch because it's
part of a Dagger injection chain into CustomizationProvider, and
CustomizationProvider is a content provider that's re-created for each user.

However, CentralSurfacesImpl ever notifies only the *original*
KeyguardViewMediator, *not* any of the new ones. I've verified that with
this change, the original KVM is still the *only* instance receiving
wallpaper updates. This is because KVM only starts collecting on
WallpaperRepository in KVM#setupLocked, which is triggered via
CoreStartable#start, which is only triggered for the original process.

Bug: 277764509
Bug: 277762009
Test: manual: Verified via logging that KeyguardViewMediator still
receives wallpaper updates from WallpaperRepository
Test: atest KeyguardViewMediatorTest

Change-Id: I14088f422c38fc95083b7f42e878e69b879f499f

Former-commit-id: 400abd9371171ca4d26ab6b8751daba66a7df87a