Avoid changing z-order of overlay layer after resizing display
The WindowToken of cutout/rounded corner overlay will be collected
in transition for sync drawing, so buildFinishTransaction will
put the restore operations in finishTransaction.

But previously the finishTransaction will always reparent the token
to the parent in window hierarchy, which doesn't consider the
override operation while creating the surface (makeSurface).

Then after the first display resize transition, the overlay
becomes flickering in next display resize because it is occluded by
the screenshot layer which has a higher z-order without including
the overlay.

Bug: 288726181
Test: Change screen resolution in Settings.
      The cutout won't have a small flickering
      (disappear in a frame).
Change-Id: Ibee8c5b8daffdaf38f4f0ce9351dd085929ccb45

Former-commit-id: 4552d3e896c884a43a836a55b1329fe304ff5ca7