Revert "Avoid changing z-order of overlay layer after resizing display"
This reverts commit 00d339a4811c41f6ba9867a0f77f422f0db07c0a [formerly 4552d3e896c884a43a836a55b1329fe304ff5ca7].

Reason for revert: b/288726181

The effect of
SurfaceControl.Builder#setParent(null);
and
SurfaceControl.Transaction#reparent(sc, null);
are different.

The former is to put at root layer, but the latter is like
to detach the layer. So the flickering was disappeared by
not showing the layer. That is unexpected. Currently there
seems to be no way to assign root layer by transaction.

Change-Id: I5b903d5dc7c30b644365b4b28048e74453e45922

Former-commit-id: f70dccdb539991bf99214adc7f93ca8f825863a8