Avoid making the user ID dir under system_ce in ATMS
The parent directory of recent_images (system_ce/<user_id>) is managed
by other system services, and isn't available until the corresponding
user is fully unlocked. Avoid trying to make that folder if that doesn't
exist.

Also fail the directory creation if it isn't a directory, and we can't
create it for recent_images.

Bug: 288146194
Test: Recent images and launch params folders are still created when
necessary.
Test: atest LaunchParamsPersisterTests

Change-Id: I7e8e3167e3aea60f6426e5ca3571e3fb1dad98e0

Former-commit-id: 312fffa4b942ae796e7051bd766dac8fd4211f75