Add empty implementation of ShadeController
Variants without shades want this, so there should be only one, and
it should be in the shade package to make updates not require topics.
Also adds some default methods to ShadeController to reduce the size
of ShadeControllerEmptyImpl and deemphasize ShadeController methods
that are too tied to implementation details of ShadeControllerImpl.

Test: presubmits are sufficient
Bug: 288867586
Change-Id: Ia7015378f85a8a68db049069b2d678b57c4390b1
Merged-In: Ia7015378f85a8a68db049069b2d678b57c4390b1

Former-commit-id: dbe43d5246fd80b270e9db3640566010d31a3dfe