Sets BUILDER_EXTRAS_OVERRIDE flag default true
Sets BUILDER_EXTRAS_OVERRIDE flag to default to true. This change allows
apps to update the extras of their NotificationBuilders (rather than
having old values override any new values set).

This is a fix to a long standing bug, but may cause undesireable
changes/breakages in Notification behavior by apps, in ways we can't
predict. As such, we'll enable it in udc-qpr-dev for 2-3 weeks in
anticipation of bug reports. The flag blocking bug has been marked as
release blocking, so if need be we will revert this in udc-qpr-dev
before release.

Bug: 280783617
Bug: 169435530
Test: Unit test coverage of the flag in both states exists. We expect
this to change behavior.

Change-Id: I1bf34b12131d63980ea69be91f5e7fba3d94c964

Former-commit-id: f9dea613a602ca4da81c1b78b9cbbcf886693874