Update the expanded view height when bar is stashed
When the expanded view is displayed in WMShell we use the bubble
bar position to determine its height. When the bubble bar is expanded
from stashed (e.g. when the first bubble is added, or when swiping up
on the stashed bubble bar), the position passed into WMShell is the
stashed position which results in an inaccurate height.

This change passes the bubble bar X and Y offsets from the bottom
corner of the screen of where the bubble bar will eventually be positioned
after it is unstashed. This allows WMShell to correctly calculate
the height of the expanded view so there is no overlap.

This change also fixes a bug observed when a bubble is added to an empty
bar, which resulted in wrong arrow position and sometimes missing expanded view.

Fixes: 290636927
Fixes: 290095535

Test: manual
      - Reboot device and add bubble
      - Observe that expanded view shows, and arrow position is correct

      - Go to Settings -> Bubbles -> toggle off and back on
      - Add bubble
      - Observe that expanded view is displayed and arrow position is correct

      - Add bubbles to bubble bar
      - Stash bubble bar
      - Swipe up on the bubble bar to expand
      - Observe that expanded view is positioned correctly

Change-Id: I9a900a2ec4a9ac72550ef94ebc8fa47e51675059

Former-commit-id: fab598746f9d1839c26eacbcec4cdf000ec76faa