[2/2] Migrating ShadeCarrierGroup mobile icons to modern status bar data pipeline
Hooks up the ShadeCarrierGroup to depend on only the new data pipeline. The UI is left mostly as-is for now, with the change being the ShadeCarrier no longer having its own version of the mobile icon layout, and the text view for the carrier name has been moved inside the ModernStatusBarView.

Bug: 288631206
Test: Manually verified no visual change with flag enabled, with both 1 and 2 sims
Test: Added/updated unit tests
Change-Id: Iaf91e5606640c5fad6e51ce3b6607e65f14d3bcb

Former-commit-id: 062fbc4578dacfbe8a04d5299bc8bb9f7f1437fa