[Status Bar] Set status bar window insets to status bar height always.
This is needed for the animation shown when you tap the ongoing call
chip in order to launch the call app. If we do *not* set the insets,
then the insets will change during the animation, which will result in
the animation being cancelled and show jank when launching the app.

This CL is improvement of ag/23875866. That CL originally caused
b/290300359, where the status bar height changed but the insets were
*not* changed, so the insets were too small. This CL also adds the
insets to the #applyHeight function, which is called when the status bar
height changes.

Bug: 283958440
Test: tap on ongoing call chip -> verify no jank during animation
Test: status bar smoke test on phone (verify status bar in portrait and
landscape, with and without ongoing call chip)
Test: status bar smoke test on tablet (same as above)
Test: status bar smoke test on foldable (same as above)
Test: regression test for b/290300359: Boot foldable device while
unfolded (no display cutout) > fold device (has display cutout now) >
open app > verify there's no white/black bar showing between the status
bar and the app

Change-Id: I35df87ef05cb75dcd551cc4899a0a9863fdcbcc7

Former-commit-id: 4b89585868f4c95b2458bb1eef77587a896751a7