[flexiglass] Resets the scene when the device goes to sleep.
When the device goes to sleep (usually following a power button click or
screen timeout), we will switch to either Gone or Lockscreen as the
current scene, depending on whether the device was unlocked or not,
respectively.

We use the ASLEEP wakefulness state as our signal instead of listening
to screen off events because the latter does not necessarily need to
trigger a scene reset, e.g. in the case of a screen turning off when
unfolding a foldable device.

Fix: 290403062
Test: new unit tests added
Test: manually verified in system UI that screen timeout moves back to
the Lockscreen scene when the device is locked, and to the Gone scene
when the device is unlocked.

Change-Id: Iee5474d45c281bbd5a76c06ba9c64de31a1d8255

Former-commit-id: 9a5b360c112f19d37c0f9208d302237f886f059b