Ignore ASHA hisyncId grouping if the device supports CSIP
If a device is added as another device's sub device, it'll be removed
from the cached device list. This makes the device can't be correctly
grouped as a CSIP set when CSIP is connected.

We should ignore ASHA hisyncId grouping and leave it to be handled as a
member of CSIP set.

Bug: 288284097
Test: atest CachedBluetoothDeviceTest
Test: atest HearingAidDeviceManagerTest
Change-Id: Ibf10198521d3186186e87f96b13475185e51e856
(cherry picked from commit ec6e7c971a4b12aea0a33737aca4ebddb9212849)

Former-commit-id: f5a03d91d60469637f7a65a22b0b1cc0f71ea986