Adding hover state to system icons in status bar and shade header
Adding StatusHoverListener which is added as onHoverListener to the status view that should be hoverable in keyguard status bar, regular status bar and shade header.
Listener adds drawable overlay on top of the view on cursor hover. The color of the drawable is:
 - keyguard: managed manually from KeyboardStatusBarView (based on theme)
 - regular status bar: dependent on dark areas
 - shade header: always light

Also adding StateFlow to SysuiDarkIconDispatcher so new consumers don't have to use regular listeners.

Test: StatusHoverListenerTest
Test: screenshot test coming soon
Fixes: 283800652
Fixes: 249859409
Change-Id: Id9c63f5d125a662d98abe5c39f6e450a516e9823

Former-commit-id: a10268d777faff23437424b151690ce78b2c77bd