Fade out UDFPS on shadeExpand & dialogProgress
When the UDFPS view refactor is enabled,
fade out UDFPS views when the shade expands to full
shade and to QS. Also, don't show the UDFPS view when
a dialog shows.

Test: atest UdfpsKeyguardInteractorTest UdfpsLockscreenViewModelTest
Test: manually enable udfps view refactor and pull down notification
shade (from the middle of LS and from the top). Launch dialogs such
as the power menu. See UDFPS lockscreen view hides + shows as expected
Bug: 278719514

Change-Id: I5cbcb217b7bd5df7a7f19cb118b843c3de0db744

Former-commit-id: 15f23551164828e4b278718114dee22e22740617