Revert "Revert "Load native GLES driver when specified.""
This reverts commit 76a149993a1d5aadd06b31de8249c301b87410dc [formerly b50ce8c91e6a7c912b18fd6f1b3a411440ada1f7].

Reason for revert: Revert the revert to reland the patch. Original patch
was reverted due to rss memory regression. Attempt to reland the patch
with a fix, the original regression seems to be caused by usage of enum.

Original commit message:
"""
Load native GLES driver when specified.

Since ANGLE and native GLES drivers can coexist, when native is
specified, the loader must load the native GLES drivers specified in
ro.hardware.egl. This patch passes this information down to the native
graphics environment so that the EGL loader can properly check.

Bug: b/283858001
Test: atest CtsAngleDeveloperOptionHostTest -c with ANGLE being default
Test: atest CtsAngleDeveloperOptionHostTest -c with native being default
"""

Bug: b/283858001
Test: atest CtsAngleDeveloperOptionHostTest -c with ANGLE being default
Test: atest CtsAngleDeveloperOptionHostTest -c with native being default
Change-Id: I9a01abe2f652a5dcfa5148badde8d7aa43c1a151

Former-commit-id: fd23411e30bcbb52438b11fae8f6f923f98f68e5