[flexiglass] Implement lockscreen bypass.
When bypass is enabled, the lock screen will be automatically dismissed
once the authentication challenge is completed. For example, completing
a biometric authentication challenge via face unlock or fingerprint
sensor can automatically bypass the lock screen.

It is deliberately not implemented as a Flow, since its state only needs
to be checked on demand.

Note: to enable this feature, the "Face auth modern arch" flag needs to
be enabled via Flag Flipper.

Fix: 290771600
Fix: 290404894
Test: new unit tests added
Test: manually verified in system UI that the lockscreen is skipped when
bypass is enabled in settings, and that it is not skipped when bypass is
disabled. To reach this setting, go to Settings > Security & privacy >
Device unlock > Face & Fingerprint unlock > Face Unlock > Skip lock
screen.

Change-Id: I393b9bcb1fed1299efaf8c426d269b4b83b36476

Former-commit-id: aec54aadc385aed0ea5c86ac4f8e9081e5e8d765