fix(non linear font scaling): date textview is cut off in qs header
Originally the qs header height is fixed to 120dp, so when larger font scaling the header height might not be enough to accommodate the subviews with given constraints, then some of the constraints are broken and the date textview bottom is cut off. Therefore, in NotificationsQSContainerController, we calculate the needed height for the subviews as the header height to prevent the ui broken.

Bug: 291168649
Test: manually - attached screenshot in bug
      atest com.android.systemui.shade
Change-Id: I99373152104ba1b378115fa12fea3ba3d7448e14

Former-commit-id: 3cae3974c5bd527ff82aefd505fde6412d60b0b8