fix(non linear font scaling): clock and battery textview might be cut off in status bar when large font scaling
Originally system icons in status bar have fixed top and bottom 8dp paddings, when larger font scaling the space might not enough for enlarged textview then the percentage textview is cut off, as in b/291168760. Therefore, as a quick solution we let the status_bar system_icons layout height be a sp dimen value like combined_qs_header and set the top/bottom paddings in system_icons to 0dp. It's to ensure the system_icons layout height is larger than the status icons height. In the future cl, we would also ensure the status bar height should be enough for accommodating the system_icons.

Bug: 291168760
Test: manually - attached screenshot in bug
Change-Id: I070bf6d5faa858acfa4927a9590c4575dc91f2fe

Former-commit-id: 4f9682e60cdffc289b55c62ffa6f6bfff6ac93b4