fix(#Magnification): flaky presubmit error on FullScreenMagnificationGestureHandlerTest
The test case testTransitToPanningState_scaleDifferenceOverThreshold_startDetecting is flaky. The main reason is that the exact persisted scale after setScale+persistScale is not expected, then it causes the ratio not passing the threshold so the flag is not set to true. Therefore, in test case we directly update the persisted scale by Settings api to ensure it's expected. In this test case our main target is to check whether the flag is set under given scale and persisted scale, so we think directly set the persisted scale is fine, and the update persisted scale logic is already verified by other test cases.

Bug: 289268089
Test: atest FullScreenMagnificationGestureHandlerTest via abtd
      presubmit
Change-Id: I5ec46ec84d1693f01b20c07f3e617b22c465da21

Former-commit-id: a9b435df01a4af87ee0f3d0b67a05357dd296e00