[flexiglass] Auto-confirm, isPatternVisible are "refreshing flow"
The upstream of the changed flows (LockPatternUtils) doesn't offer a way to listen for changes to these configs which means
that stale/old values are cached forever in the flows. This CL changes
the flows to special "refreshing flows" which (1) provide an initial
value (2) refresh in the background (3) refresh each time the selected
user changes or when a new downstream subscriber is added.

Included there are some minor changes to switch from
SharingStarted.Eagerly to SharingStarted.WhileSubscribed (so the UI
coming to the foreground counts as a new subscriber each time) and a bug
fix to PatternBouncer to make it not draw the lines if isPatternVisible
is false.

Bug: 280883900
Test: unit tests all pass still
Test: manually verified that the bouncer correctly switches between
autoconfirm enabled and autoconfirm disabled when I change it in the
settings. Same for pattern visible.

Change-Id: Ia8a3fca84437d2ddea18ebf1e4ad7ac8fe1c4f31

Former-commit-id: 35ee1a7b2f6ea72ce9bf209a6e1b06781ff2a1f9