Landscape PIN bouncer: refactor of portrait layout
Portrait layout moved to a separate file and is now included in keyguard_pin_view. This is done in preparation for the implementation of a landscape layout for the pin view where the portrait layout will be referenced multiple times, so the refactor is to avoid code duplication in the future (as the separate portrait layout file can be included multiple times, avoiding duplication).

Bug: 293427037
Test: KeyguardSecurityViewFlipperControllerTest
Change-Id: I5d7c8c02f6c448755cb5726fd9d59524d9f2719a

Former-commit-id: e11fb91ddb6da156de2fd1c81969a4987aecf1f9