Using performHapticFeedback on NotificationPanelViewController
Migration towards the new one-way API that can trigger haptic feedback from the UI thread. Calls to the vibrate method of the VibratorHelper are replaced by performHapticFeedback with haptic constants that represent interfactions. A GESTURE_START vibration is delivered upon swipe down events and a REJECT vibration when a second tap is required due to falsing. The migration is controlled by a feature flag at the moment.

Test: Vibration calls were tested in the NotificationPanelViewControllerWithCoroutinesTest class. Manual verification of vibration was performed in the event of swiping down the notification panel. The vibrations were verified when the feature flag was on and off.
Bug: 245528624
Change-Id: I55c3e7a0627358cf5ac0c6e221a6e8bb9a8f8404

Former-commit-id: f966e40a2ee614f713a5f648f3cc0f3859e4bee4