CSD: Add possibility to select the BT audio device type
Extended the AdiDeviceState with properties to store the bt audio device
type. As part of the AdiDeviceState these properties will be persisted
with the help of the Secure Settings.

In case a bt device was categorized as headphone/non-headphone device we
send this information to the audioserver which will start/stop the sound
dose computation.

Test: manually changing bt device type
Bug: 287011781

Merged-In: Ia54e141a09b7c95f43b59bab1445eb08890383b3
Change-Id: Ia54e141a09b7c95f43b59bab1445eb08890383b3

Former-commit-id: 0878d80dd8f7b317113c44f2b424873c9e4a9641