Begin disconnecting KeyguardStatusViewController from NPVC
As the status view will (eventually) no longer be in the shade, begin
disconnecting it from NotificationPanelViewController piece by
piece. First, remove dozeTimeTick and onScreenTurningOn calls from
NPVC. This rewires the flow of events but does not change any functionality.

Also, refactor burn in to use KeyguardRepository as the source of
truth for dozeTimeTick information

Bug: 288242803
Test: atest KeyguardStatusViewControllerTest
KeyguardStatusViewControllerWithCoroutinesTest BurnInInteractorTest
UdfpsFingerprintViewModelTest UdfpsKeyguardInteractorTest

Change-Id: Ic9061bd67eb78995eb38ff88315dd9e0c4899e6b

Former-commit-id: a3933a7d147ec196ef897d9b3ccb9147be232dbb