Fix user switcher flicker.
User switcher flickers when bouncer shows scrimmed. This is because when
the bouncer hides, the alpha of the view is considered to be 1f. When we
are showing the bouncer, we are asynchronously inflating the view and
waiting for this operation to complete to start the appear animation. In
order to fix this, we reset the view before making this async call to
ensure that the view alpha is 0f. It will be set to 1f in #appear.

Fixes: 290355544
Test: test pattern, pin, password back button and success.
Change-Id: I1680d66336b5b6d5d1065c85ac4b09ca56e12d7a
Merged-In: I1680d66336b5b6d5d1065c85ac4b09ca56e12d7a
(cherry picked from commit 256d5fdd4e6787b15c1dff42ad5cb167e59644b7 [formerly dc072507d0f52deb7e23ca5618816dbf18952204])

Former-commit-id: 334ebcea1bd4c211b288878881610c5191cb30b3