[flexiglass] Make the window-view full-screen.
Before this CL, the window-view for the scene container (our root view)
was not full-screen; it was getting cut off by the top and bottom window
insets (status bar and nav bar). With this CL, that's no longer the case
and Flexiglass is now full-screen behind the status bar and nav bar.

The approach was to identify the pieces of the legacy
NotificationShadeWindowView and move them to the common base class of
that and of the newer SceneWindowRootView, which is WindowRootView.

Fix: 290984671
Test: manually verified that the bouncer scene extends to the top and
bottom of the display when Flexiglass is turned on but also when it's
turned off (no harm done).

Change-Id: I05cc4202bae1530506f6f0c90afdcda03ad7d020
Merged-In: I05cc4202bae1530506f6f0c90afdcda03ad7d020

Former-commit-id: 8f20646bb54958906499ca2eb6835c2a8521d57f