fix(non linear font scaling): status bar mobile/wifi icons size not correct
The mobile/wifi signal icon sizes are mis-set to 13sp, originally they should match the status_bar_system_icon_size. Therefore, we correct the dimen value so that when default font scaling the icon sizes whould be same as the original size.

Besides, The mobile type/roaming icon sizes are mis-set to same as mobile signal icon size, while originally they're WRAP_CONTENT so should be same as the drawable size. Therefore, We adjust the type/roaming icon sizes to match the corresponding drawable icon size. It's to ensure then the icons is still scalable and the sizes are same as original when default font scaling.

Bug: 291899494
Bug: 293196179
Test: manually - attach video/screenshots
Change-Id: Ia601968173c71687e515b8d81f2e7980e48568ed

Former-commit-id: d21590e080fcd553f2be91295f3506f77e2760e5