Fix keyboard showing after dialog dismissed
When dialog for creating user was dismissed with keyboard being displayed, the keyboard stayed on the screen and could not be released due to incorrect list of methods called on dialog dismissal. This CL topic fixes onDismiss() behaviour of this dialog. Videos of current behaviour and behaviour after fixed are attached to the bug.

Bug: 291865725
Test: atest CreateUserDialogControllerTest
Change-Id: I732d47039b317c14a04c2324589c023f7bcdc6cc

Former-commit-id: c6f198750de04374862ff8e1221a61768b843694