Do not skip finishing async rotation by a non-rotation transition
Otherwise when completing a fixed rotation without display rotation
change but there is a display size change transition, the animation
of async rotation won't be stopped. Which may lead to some bars
window invisible because it expected there will be a rotation change
to trigger it to fade-in the bars but the display didn't rotate.

Bug: 275566237
Test: atest ShowImeWhileEnteringOverviewTest
      TransitionTests#testFinishRotationControllerWithFixedRotation
Change-Id: I73da710fbb4fa1d6c20a9aed6da80c3d9009230e

Former-commit-id: 9f858de2d07feed4fec3f4fba35a413c423f6813