Using performHapticFeedback on BiometricUnlockController
Moving towards new haptics API that triggers haptics directly from the UI thread. The authentication success and authentication error trigger signals of CONFIRM and REJECT, respectively. This is while using the performHapticFeedback API from the VibratorHelper. The method requires a view and the view from the ViewRootImpl accessed from the keyguard controller was used. The changes are controlled by a feature flag.

Test: Tests in BiometricsUnlockControllerTest pass with additional tests that involve oneway haptics on authentication success and failure.
Bug: 245528624
Change-Id: I3b4e83d616e1c8ad5d64a4926975bd08379cd569

Former-commit-id: e7ff086f5a1470deb6c794c7eeb513f1abaccd2f