Redirect shelf height requests into keep clear area registration
Looks like there's a RecentsView code that can set it to non-zero
value and it will stay that way until it gets reset on rotation.

Hence guarding it on the PiP size from being executed is a sure way
to prevent unwanted calls to that logic while the flag is on.

Previous site calling this was in QuickstepLauncher and was manually
guarded from being called.

Bug: 270677470
Test: atest WMShellFlickerTests
Test: presubmit passes with no flakes in PiP area

Change-Id: Ib73e1cbb4c90615486372085eef9a8fdba41e960

Former-commit-id: 00a2de432011f19e8a7f7fca03adb4eb57697d4a