Only create InsetsSourceConsumer for InsetsSourceControl
Previously, the insets source consumer is only created if the insets
source exists. However, if a control target is not below an insets
source, the target won't receive the source, and the consumer won't be
created, and the insets cannot be controlled.

Also, InsetsSourceConsumer only works when the control exists. So we
don't have to create redundant consumers for types not controllable.

This CL only creates the insets source consumer for the control, not for
the source. This fixes the z-order problem.

Fix: 290884763
Bug: 234093736
Test: atest InsetsControllerTest InsetsSourceConsumerTest
      ImeInsetsSourceConsumerTest
Test: Watch a video in Google TV. Long press an icon on taskbar before
      taskbar fades out. See if taskbar is visible.
Change-Id: I16a854d52f46d36899c826426afc5ebbcf47b586

Former-commit-id: aa27825aff00b69d0a77db6143c52acfa22a41a6