Allow system installer to deny update ownership
This change introduces a means for system installers with the
INSTALL_PACKAGES or INSTALL_PACKAGE_UPDATES permissions to provide a
list of packages that are opted out of update ownership. The list may be
updated via updates to the package that provides it. Packages that
provide such a list will be unable to request update ownership of any
packages.

Bug: 293210067
Test: Manual; tested various scenarios per b/293210067
Change-Id: I544f23dcfa9a7eefc49f1d72fb612a11ebea98bc

Former-commit-id: 57f153142757c5c7867c480c421eb09595a51a26