Release TaskView after the task has been removed in Bubbles
When the bubble expanded view gets cleaned up, we could still be
animating task view / the task removal, so instead of releasing
TaskView there (which clears the task from TaskViewTransitions which
caused issues with animating the removal), we release the TaskView
in the callback we get for the task removal.

Additionally fix up the code in TaskViewTaskController where
onTaskRemovalStarted is called so that if TaskView#release is called
from onTaskRemoval, then the taskLeash, taskToken, taskInfo won't be
null.

Test: atest TaskViewTest
Test: manual - expand a android messages bubble
             - tap on the manage menu > dismiss bubble
             => observe that the animation is normal and the surface
                isn't left on screen (repeat test a couple of times,
                isn't a 100% repro).
Bug: 293399673
Change-Id: I3b29a7c358f7fe9f79aa99bdd00a8e3ec35cfd69

Former-commit-id: d72d7bf73583013fdf582b51af3901a58dfe4fe3