Refine the performance of updating the insets hint
Previously, the insets hint was checked and updated every time in
InsetsSourceProvider#onPostLayout via setServerVisible. Most of them
were just checking but not updating. That still caused the regression of
performance.

This CL flags the insets hint as stale when the source frame or the
window container bounds is changed, and only updates the insets hints
when someone calls getInsetsHint().

Fix: 292177371
Test: atest InsetsSourceProviderTest
Test: Switch to 3-button navigation bar and reboot the device.
      Make sure the back button is not shown on the lockscreen.
Change-Id: I59e21581d7ca136d4e4eb8c1a442fc74a92914b7

Former-commit-id: af1c7cfd1e290f2181a46435bcd7d9735c6a89fd