Modify the default position of bubbles on tablet to be centered
If we don't have a previous bubble position we use a default spot.

On phone we use an offset from the top of the device and place
bubbles based on that. This was also being used for tablet. This
CL modifies the behavior so that we'll center bubbles along the edge
for the default position on tablet. This makes them a little more
reachable for the larger screen size.

Bug: 283910436
Test: atest BubblePositionerTest
Change-Id: Ifc1a889809ea1be53e71c920e86e923b3cf98805

Former-commit-id: 33319db63e4cd624f3e7d76e7a149a0461eb8877