Do not finish lockscreen for multiple sims.
Currently, in the logic, if the lockscreen is disabled, we finish
lockscreen instead of proceeding ot the second sim. I have removed this
logic. We are expecting the security mode to be SimPin/SimPuk if we are
expecting an additional sim. We also do not want to finish if the
security mode is supposed to be secure.

Fixes: 282107332
Test: added a unit test.
Test: test 2 esim cards and restart with none security.
Test: test 2 esim cards and restart with swipe security.
Test: test 2 esim cards and restart with pin security.
Test: test 2 esim cards and restart with pin security from sim-puk to sim-
pin.

Change-Id: I519fcb995a5915100bd97faaa9618d573c64bcc2

Former-commit-id: 0f0a5bd651d6184ed6c45e1cd8069bb8f09559d1