Allow alternateBouncer if at least once source has ui available
When the device enters the keyguard in landscape, the last UDFPS
view is removed and a new one is added. The timing of events
was detaching the previous view after the new view was attached,
which was causing setAlternateBouncerUiAvailable to end up in the
wrong "false" state. Guard against this by tracking the source
of who is saying the alternateBouncerUiIsAvaiable.

Fixes: 293979249
Test: on the homescreen or in app, get into landscape, then
go the lockscreen and trigger the alternate bouncer (tap
on a notification or press the ... on the volume UI). Observe
that the alternate bouncer shows if UDFPS or SFPS is enrolled.
Test: atest AlternateBouncerInteractorTest

Change-Id: I16edc54c0d0ab693c2425ebcc136dbd2ef0ffc56

Former-commit-id: 5d49248b10e6e4d47d3d4efece1cc23d3c7010ff