[flexiglass] Use transitionState instead of currentScene.
There's often a mismatch between the currentScene and the UI because
currentScene drives the UI (when programmatic scene changes are
requested) but also the UI updates currentScene (when a
user action completes).

This CL simplifies the API surface such that external consumers of the
Flexiglass API can use the transitionState, driven directly from the UI
to know about transition pairs and progress of transitions (vs. idle).

Unfortunately, I did have to keep currentScene (renamed to
"desiredScene" to better capture its role in the system) so it can be
served downstream to the UI for servicing programmatic scene changes.

In addition, many of the existing consumers of Flexiglass were updated
to no longer rely on currentScene as it would (a) often mismatch what
the UI is showing and (b) not provide enough context to know about the
state of transitions.

Fix: 294220005
Test: Unit tests updated
Test: Manually verified with auth method None, Swipe, and Pattern - on,
unlock, lock, swipe down to reveal shade and QS, swipe up to reveal
bouncer, back gesture to return to lockscreen.

Change-Id: I3de051f7c0331f39679c15ec0c099cd574b68935

Former-commit-id: 60eef3c47ab2c30b07d2353cb7cee5113bf161bd