Invalidate stylus handwriting availability cache
In [1] we introduced caching of attributes that determines if stylus
handwriting is available. The strategy used to invalidate cache on IME unbind.
However, a stale value of
InputMethodInfo#supportsStylusHandwriting can be cached if IMM#isStylusHandwritingAvailable() previous IME unbinds and the next one isn't yet connected.
A better approach would be to cache the value on the new IME service connection.

[1]: I6d89d82589e4946e07bfbc4b97811fe4537b1a9b

Bug: 288618291
Bug: 294162601

Test: atest InputMethodServiceLifecycleTest

Change-Id: I35013b6194a67656aefca004e69cfc64f12c1e88
Merged-In: I35013b6194a67656aefca004e69cfc64f12c1e88

Former-commit-id: 16b6b5109a1b1023d859bf9a46040387f24f3c20