Use breakstrategy and hypenation properties in ImageFloatingTextView
Hypenation and break strategy are 2 important configuration parameters for PrecomputedText.
We have to use the same hypenation and break strategy in both PrecomputedText and StaticLayout creation to make use of the power of the PrecomputedText.
Instead of using hard coded in StaticLayout creation, this CL sets these parameters in the constructor and use the properties.

Bug: 289250881
Test: run ImageFloatingTextViewTest and Manual. Post a bigtext notification with this change and without and compare the result. Result should be the same.
Change-Id: I65576ed822f32622d44d30a689531a9b8fa2ba07

Former-commit-id: 7e440e0eceaed0c1667ee63abd15c0deb4269687