Camera: Avoid caching extension service connections
Under certain condition the camera extension proxy service can
be frozen along with the client process. Depending on the
lifecycle of the client the service may not get unfrozen
afterwards resulting in a still valid service connection
but invalid/failing binder.
To avoid this ensure that the extension service connection
is destroyed as soon as there are no more client sessions and
calls.

Bug: 287618467
Test: Manual using Camera2Extension app,
Camera CTS

Change-Id: I0cd62c067f575846ef01684958a844f180b24c8d

Former-commit-id: 994a11fbb861a512de6b37c1df4a262a30ad5393