Dismiss freeform tasks occluding keyguard in folded transitions
This change is an enhacement of
Idac9e012d25c50b73eba8afe9088d77043c68a85 where the "dismiss
freeform" behavior when a task occludes the Keyguard was moved from
WM Core to WM Shell. That changed looked for transitions requested with
transit type OCCLUDE_KEYGUARD as a signal to change windowing mode
to fullscreen. However, it didn't cover the cases where a transition
is already collecting such as opening a showWhenLocked activity
directly into freeform mode over the keyguard, which ends up being of
transit type OPEN.

To cover cases like that, two changes are made:
 1) Add the transition flags to TransitionRequestInfo and use the flags
to check whether there's an occluding task that needs to be dismissed.
This should cover cases where the transit type is not KEYGUARD_OCCLUDE
but the keyguard visibility change happens early enough that the flag
is added to the collecting transition before the transition request is
sent to Shell.
 2) Add the dismiss behavior to the finish WCT in
KeyguardTransitionHandler to cover cases where the keyguard visibility
update happens after the transition request has already been sent to
Shell but before the transition is ready.

Bug: 261765739
Bug: 293219242
Test: atest ActivityVisibilityTests
Change-Id: I4c2c3348dda94c58d29e9545243b49eb4f166743

Former-commit-id: eff3514537836938284b11796eb5ee0e4daeb998