[Sb refactor] Move icon to the interactor
MobileIconViewModel was doing a little bit too much business logic to
calculate the correct SignalDrawble icon for the view. This logic will
also be needed elsewhere, so it makes sense to move the calculation to
the Interactor.

It is still up to the ViewModel to map from the icon model into actual
UI via the binder, so this shouldn't change anything other than where to
look for the model.

Test: MobileIconInteractorTest
Test: MobileIconViewModelTest
Bug: 265342892
Change-Id: I85bad6310485a4ea57c4b32e3fd632868779212f

Former-commit-id: a388ad63b63ebe31c299d5134f7b227244f7b3ae