[Autofill PCC] Fix authentication logic for prefer PCC
Fix the authentication filling logic when prefer PCC is enabled.
Previous internal logic could create multiple datasets after authenticaiton.
This problem was visible if the PCC didn't detect any of the filling fields.
The bug is fixed by merging the datasets with the detected fields, as to
create a single dataset.

Bug: 293313411
Fixes: 291194194
Test: atest CtsAutoFillServiceTestCases
Change-Id: I5dd50c6639a96a046213e00fddf0b15c8bb3d5e3

Former-commit-id: 040121465b6c4ab811b9bede437ec261d5285029