Update process animating state with different reasons
Since shell transition is enabled, it is more often to invoke
setRunningRemoteAnimation for SystemUI. So there may a race
race that SystemUI has been boosted for entering AOD, but the
state is canceled by a finished transition.

So now with individual animating reasons, the transition will
only remove its corresponding reason. If other reasons still exist,
the process can keep the animating state for a higher priority.

Bug: 295116622
Test: atest WindowProcessControllerTests#testSetAnimatingReason
Change-Id: Id326518a0e2b2b9375015e55b5ce3983a9e801e7

Former-commit-id: a2b2d7391d5e73bd02bb8bc3504fc785b3f71606