Create connected display presentation
This creates a parallel implementatino of the connected display
presentation following UX directions, showing a single clock, centered
in the screen.

This is flagged by "ENABLE_CLOCK_KEYGUARD_PRESENTATION" flag. When the
flag is off, the previous implementation is used.

Note that only the default clock will appear as centered for now.
Regarding custom clocks, they will be addressed in a follow up, with
screenshot tests.
+ Extracted KeyguardPresentation to use assisted factory

Test: KeyguardClockSwitchControllerTest, KeyguardClockSwitchScreenshotTest
Bug: 278982856
Change-Id: I26b4800bd7c92c167a8b53e3c80d60be9224144a

Former-commit-id: 212dc4a1931a2b50707aa25480e30e5d6ff3bc98