[Sb refactor] switch broadcast off of BroadcastDispatcher
`ACTION_SERVICE_PROVIDERS_UPDATED` is a sticky broadcast, so we avoid
using the BroadcastDispatcher so that we never miss the message due to
sticky behavior.

Also fix the incorrect intent extra, as this broadcast uses
`EXTRA_SUBSCRIPTION_INDEX` rather than the subscription_id field common
in some other telephony broadcasts.

Test: MobileConnectionRepositoryTest
Bug: 291321279
Change-Id: I513abe7904a48cd5022ffeed15ebb7933ad5953e

Former-commit-id: a3c24fe04ca723e60729ad3556c513e899a435f7