Remove requirement for view attachment to refresh date format
Since the date format is not read until the next call to updateClock, this simply eliminates the possibility of ignoring a timezone change due to lack of a handler. Also adds some logging in case this is not the culprit.

Bug: 284541914
Test: Verified timezone change is handled correctly
Change-Id: Ia733f72f0b2b9ee0389c7114a50b362a0e3e324d
Merged-In: Ia733f72f0b2b9ee0389c7114a50b362a0e3e324d

Former-commit-id: 3468f310619c4ac9dba1539e93bd6f7cbbee3e53