Add interactors for KeyEvents
This is in preparation to swap over
to using the KeyEventInteractor instead of
calling methods on CentralSurfaces. However, this is NOT and ideal
way for the interactors themselves to be structured. In the future
the interactors should be updated to reflect true modern
architecture patterns.

The logic in these interactors match logic from
NotificationShadeWindowViewController and CentralSurfacesImpl.

Test: atest KeyEventInteractorTest
Test: atest KeyguardKeyEventInteractorTest
Bug: 295927655
Change-Id: Id5c5659a4a19a0b0feb45d6ee391930c2063a9b1

Former-commit-id: ce5fbaed64a797adb96f9bca8f40e57f0573a621