Only set SYNC flag if >1 track
While technically it's true that a transition is sync
if another is running in the same track, tracks already
imply sync between their constituents, so it's redundant.
The player also interprets SYNC as "stop everything
immediately" which isn't actually useful unless there are
multiple active tracks.

So, only set the sync flag when there are >1 tracks.

Bug: 296242449
Test: atest TransitionTests#testNoSyncFlagIfOneTrack
Change-Id: I61d97542825627beba4dbfa77f1cff0cf1e86ab8

Former-commit-id: 7b3c8ddea7d6e725d0a1b64fdeae35f940836974