Remove host activity association when removing EmbeddedWindow
The common steps to destroy an activity is first to remove its window
and then the activity record. So if EmbeddedWindow is removed when
removing the window, onActivityRemoved won't be able to match.

If an ActivityRecord is removed, its window will also be removed.
So it can simply clean the host records when removing EmbeddedWindow.

Bug: 296535363
Test: Launch "Wallpaper & style" from home several times.
  > adb shell dumpsys activity p | grep -A2 "Remote Activities"
  The output should only show one record per owner process.
  And then remove the task from recents.
  > adb shell dumpsys activity p | grep "Remote Activities"
  The output should be empty.
Change-Id: Iebe96417ed5be1832995ad7649408c28e6c0f58c

Former-commit-id: 352aa3d13c8566c114b3fdd4233927a27889c42f