Fix bouncer always being on the left side
The keyguard position value was being set to 0 as the initial/default
value, which would be passed on to the view/ui layer on start, which
would then move the bouncer to the left.

At the moment there is code in the view layer that stores the last used
position (left vs right), and by having the 0 value always coming from
the data layer, it would override this setting.
A better refactor in the future, would probably be to have this logic
of storing/retrieving the setting, in the data layer itself, but would
like to leave that fix/improvement for a later stage.
This stage is just to fix the bug.

Test: Unit tests in this CL
Test: Manually - By moving the bouncer and then restarting the device
Fixes: 296845861
Change-Id: Ib5c508ef0c589c173b3e10280b0fe5156274e7e0

Former-commit-id: dbbbedb352eca602b27347131e1513d344d9c622