Keep insets leash position before display transaction applies
This is similar to keepAppearanceInPreviousRotation. Not only the
content drawn by the window, the insets leash position also needs
to keep the original rotation. Otherwise because the insets leash
is the parent of window, the window may be put at a rotated position
before the display projection applies the new rotation.

Fix: 293395004
Test: atest TransitionTests#testAppTransitionWithRotationChange
Test: Use gesture navigation to swipe from portrait home to an
      immersive landscape activity. The navigation bar is not
      flickering at the center of screen.
Change-Id: I1def139a19831c03879199e7f095750f0cc23cb3

Former-commit-id: 93ceb9983199014c86ce38c5afcc04727d6828d1