[wm] schedule a traversal when animation ends
When a window animation runs, we boost the frame rate to
max, overriding any framerate that the app sets. We were missing
an update to window framerate which meant after a screen
rotation, the app would be stuck at  the max framerate.
Fix this by scheduling a traversal when the animation ends.

Bug: 284271661
Test: see steps in bug
Change-Id: I9f0ff7153078d005a34db9c2d98bdbc9d36e9657

Former-commit-id: 3468cd94064c4924e9c130726e3e7fca37ebbb53