Update assertion for ContentRecordingController test
Added this assertion in ag/24265583, assuming that updateRecording could have only been called once, but it can be called previously in the test setup when the DisplayContent is created (if a race conidition is satified, ergo causing flakes). Updating assertion accordingly to stop failures when updateRecording is called more than once.

Bug: 295887403
Test: atest WmTests:ContentRecordingControllerTests#testSetContentRecordingSessionLocked_updateSession_noLongerWaiting_accepted
Change-Id: Ife46df61c4796da395216021eec75cec44a75f75

Former-commit-id: cdb14a15255b2f880d1eb7e51e3d102d752cced3