Fix IME hidden nav bar bottom insets
In some apps, when the content would fill the screen, with the code path
exercised for hiding the IME navigation bar and with the IME in floating
mode, a larger bottom inset size would be sent to the app.

This is due to an older assumption in DisplayPolicy#getImeSourceFrameProvider
that is no longer applicable now that we have the IME navigation bar in
the IME process.

This also fixes DisplayPolicyTests testImeMinimalSourceFrame and
testImeInsetsGivenContentFrame which were setting values on displayInfo
but not using these consistently, which would sometimes conflict with
the real display height and cause the test to wrongly pass.

Test: install Gboard canary, open any app with a large scroll view
  (e.g. Chrome omnibox), launch Gboard in floating mode, observe how app
  content is displayed around the bottom of the screen (not cut off)
Bug: 298162496
Change-Id: Ibda9e18182cea0860d88b43fde4579810b89463b

Former-commit-id: a4229cc58c6ef8d8dd9cd64c3aba73209edf7196