Add forciblyShownTypes for privileged window to show insets forcibly
It could be done by applying PRIVATE_FLAG_STATUS_FORCE_SHOW_NAVIGATION
or PRIVATE_FLAG_FORCE_SHOW_STATUS_BAR, but the caller must be
notification shade of status bar window. This CL enables privileged
window to achieve that, as long as the window has the permission to
provide insets. This CL creates a field for the purpose, and removes
those private flags.

Bug: 278078311
Test: atest InsetsPolicyTest
Test: Receive a heads up notification from Chat in an immersive app, and
      see if status bar is visible.
Test: Expand notification shade in an immersive app, and see if
      navigation bar is visible.
Change-Id: I3fd20435a282c39f6f94e389779fe6c7761cfde2

Former-commit-id: 0cb3e5c499534372270118b26224bf240a6b0cc3