Close shade when opening MediaProjectionAppSelectorActivity
When shade is open and media projection dialog starts MediaProjectionAppSelectorActivity it will be below shade and not visible.
We want to close it when that happens.

Using StatusBarManager because using standard SysUI components (like ShadeController) would not work in secondary SysUI process when using another user (see b/298099682 for more info)..

Fixes: 285291564
Test: MediaProjectionPermissionTest#startSingleTask_shadeWasOpen_closesShade
Change-Id: I4c71d58c6c0cb8eb2a670b7059be48169254f354

Former-commit-id: 543f6f56d97773da7e674db8a3a741e521c25685